+++ src/view/list-view.js
@@ -1,23 +1,11 @@
-import { createElement } from '../render.js';
+import AbstractView from '../framework/view/abstract-view.js';
 
 function createListTemplate() {
   return '<ul class="trip-events__list"></ul>';
 }
 
-export default class ListView {
-  getTemplate() {
+export default class ListView extends AbstractView {
+  get template() {
     return createListTemplate();
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
   }
 }
