+++ src/view/new-form-view.js
@@ -1,136 +1,154 @@
-import { createElement } from '../render.js';
-import { humanizeTimeEdit } from '../utils.js';
-import { TYPES } from '../const.js';
+import AbstractView from '../framework/view/abstract-view.js';
 
-function createNewPointTemplate(point, destinations, offers) {
-  const { basePrice, dateTo, dateFrom, type } = point;
-  const pointDestination = destinations.find((dest) => point.destination === dest.id);
-  const { pictures } = pointDestination;
-  const typeOffers = offers.find((offer) => offer.type === type).offers;
-  // const pointOffers = typeOffers.filter((off) => point.offers.includes(off.id));
-  const dateStart = humanizeTimeEdit(dateFrom);
-  const dateEnd = humanizeTimeEdit(dateTo);
-  const createImageList = (arr) => arr.map((item) => `<img class="event__photo" src="${item.src}" alt="${item.description}">`).join('');
-  const createPicturesList = createImageList(pictures);
-
-  function createOffersTemplate() {
-    const isChecked = (offer) => point.offers.includes(offer.id) ? 'checked' : '';
-    return typeOffers.map((offer) =>
-      `<div class="event__offer-selector">
-    <input class="event__offer-checkbox  visually-hidden"
-      id = "event-offer-${type}-${point.id}"
-      type = "checkbox"
-      name = "event-offer-${type}" ${isChecked(offer)}>
-    <label class="event__offer-label" for="event-offer-${offer.id}-${point.id}">
-      <span class="event__offer-title">${offer.title}</span>
-      &plus;&euro;&nbsp;
-      <span class="event__offer-price">${offer.price}</span>
-    </label>
-    </input>
-    </div>`).join('');
-  }
-
-  function createEventTypeItemTemplate() {
-    return TYPES.map((pointType) =>
-      `<div class="event__type-item">
-      <input class="event__type-input  visually-hidden"
-        id="event-type-${pointType}-${point.id}"
-        type="radio"
-        name="event-type"
-        value="${pointType}"
-        >
-      <label class="event__type-label  event__type-label--${pointType}" for="event-type-${pointType}-${point.id}">${pointType.charAt(0).toUpperCase().concat(pointType.slice(1))}</label>
-    </div>`).join('');
-  }
-
-  return (
-    `
+function createEditPointTemplate(point) {
+  return /*html*/`
   <li class="trip-events__item">
   <form class="event event--edit" action="#" method="post">
     <header class="event__header">
       <div class="event__type-wrapper">
         <label class="event__type  event__type-btn" for="event-type-toggle-${point.id}">
           <span class="visually-hidden">Choose event type</span>
-          <img class="event__type-icon" width="17" height="17" src="img/icons/${type}.png" alt="Event type icon">
+          <img class="event__type-icon" width="17" height="17" src="img/icons/flight.png" alt="Event type icon">
         </label>
         <input class="event__type-toggle  visually-hidden" id="event-type-toggle-${point.id}" type="checkbox">
         <div class="event__type-list">
           <fieldset class="event__type-group">
             <legend class="visually-hidden">Event type</legend>
-           ${createEventTypeItemTemplate()}
+            <div class="event__type-item">
+              <input id="event-type-taxi-${point.id}" class="event__type-input  visually-hidden" type="radio" name="event-type" value="taxi">
+              <label class="event__type-label  event__type-label--taxi" for="event-type-taxi-${point.id}">Taxi</label>
+            </div>
+            <div class="event__type-item">
+              <input id="event-type-bus-${point.id}" class="event__type-input  visually-hidden" type="radio" name="event-type" value="bus">
+              <label class="event__type-label  event__type-label--bus" for="event-type-bus-${point.id}">Bus</label>
+            </div>
+            <div class="event__type-item">
+              <input id="event-type-train-${point.id}" class="event__type-input  visually-hidden" type="radio" name="event-type" value="train">
+              <label class="event__type-label  event__type-label--train" for="event-type-train-${point.id}">Train</label>
+            </div>
+            <div class="event__type-item">
+              <input id="event-type-ship-${point.id}" class="event__type-input  visually-hidden" type="radio" name="event-type" value="ship">
+              <label class="event__type-label  event__type-label--ship" for="event-type-ship-${point.id}">Ship</label>
+            </div>
+            <div class="event__type-item">
+              <input id="event-type-drive-${point.id}" class="event__type-input  visually-hidden" type="radio" name="event-type" value="drive">
+              <label class="event__type-label  event__type-label--drive" for="event-type-drive-${point.id}">Drive</label>
+            </div>
+            <div class="event__type-item">
+              <input id="event-type-flight-${point.id}" class="event__type-input  visually-hidden" type="radio" name="event-type" value="flight" checked>
+              <label class="event__type-label  event__type-label--flight" for="event-type-flight-${point.id}">Flight</label>
+            </div>
+            <div class="event__type-item">
+              <input id="event-type-check-in-${point.id}" class="event__type-input  visually-hidden" type="radio" name="event-type" value="check-in">
+              <label class="event__type-label  event__type-label--check-in" for="event-type-check-in-${point.id}">Check-in</label>
+            </div>
+            <div class="event__type-item">
+              <input id="event-type-sightseeing-${point.id}" class="event__type-input  visually-hidden" type="radio" name="event-type" value="sightseeing">
+              <label class="event__type-label  event__type-label--sightseeing" for="event-type-sightseeing-${point.id}">Sightseeing</label>
+            </div>
+            <div class="event__type-item">
+              <input id="event-type-restaurant-${point.id}" class="event__type-input  visually-hidden" type="radio" name="event-type" value="restaurant">
+              <label class="event__type-label  event__type-label--restaurant" for="event-type-restaurant-${point.id}">Restaurant</label>
+            </div>
           </fieldset>
         </div>
       </div>
       <div class="event__field-group  event__field-group--destination">
         <label class="event__label  event__type-output" for="event-destination-${point.id}">
-        ${type}
+          Flight
         </label>
-        <input class="event__input  event__input--destination" id="event-destination-${point.id}" type="text" name="event-destination" value="${pointDestination.name}" list="destination-list-${point.id}">
+        <input class="event__input  event__input--destination" id="event-destination-${point.id}" type="text" name="event-destination" value="Chamonix" list="destination-list-${point.id}">
         <datalist id="destination-list-${point.id}">
-          <option value="${pointDestination.name}">${pointDestination.name}</option>
-          <option value="${pointDestination.name}">${pointDestination.name}</option>
-          <option value="${pointDestination.name}">${pointDestination.name}</option>
+          <option value="Amsterdam"></option>
+          <option value="Geneva"></option>
+          <option value="Chamonix"></option>
         </datalist>
       </div>
       <div class="event__field-group  event__field-group--time">
         <label class="visually-hidden" for="event-start-time-${point.id}">From</label>
-        <input class="event__input  event__input--time" id="event-start-time-${point.id}" type="text" name="event-start-time" value="${dateStart}">
+        <input class="event__input  event__input--time" id="event-start-time-${point.id}" type="text" name="event-start-time" value="18/03/19 12:25">
         &mdash;
         <label class="visually-hidden" for="event-end-time-${point.id}">To</label>
-        <input class="event__input  event__input--time" id="event-end-time-${point.id}" type="text" name="event-end-time" value="${dateEnd}">
+        <input class="event__input  event__input--time" id="event-end-time-${point.id}" type="text" name="event-end-time" value="18/03/19 13:35">
       </div>
       <div class="event__field-group  event__field-group--price">
         <label class="event__label" for="event-price-${point.id}">
           <span class="visually-hidden">Price</span>
           &euro;
         </label>
-        <input class="event__input  event__input--price" id="event-price-${point.id}" type="text" name="event-price" value="${basePrice}">
+        <input class="event__input  event__input--price" id="event-price-${point.id}" type="text" name="event-price" value="160">
       </div>
       <button class="event__save-btn  btn  btn--blue" type="submit">Save</button>
-      <button class="event__reset-btn" type="reset">Cancel</button>
+      <button class="event__reset-btn" type="reset">Delete</button>
+      <button class="event__rollup-btn" type="button">
+        <span class="visually-hidden">Open event</span>
+      </button>
     </header>
     <section class="event__details">
       <section class="event__section  event__section--offers">
         <h3 class="event__section-title  event__section-title--offers">Offers</h3>
         <div class="event__available-offers">
-        ${createOffersTemplate()}
+          <div class="event__offer-selector">
+            <input class="event__offer-checkbox  visually-hidden" id="event-offer-luggage-${point.id}" type="checkbox" name="event-offer-luggage" checked>
+            <label class="event__offer-label" for="event-offer-luggage-${point.id}">
+              <span class="event__offer-title">Add luggage</span>
+              &plus;&euro;&nbsp;
+              <span class="event__offer-price">50</span>
+            </label>
+          </div>
+          <div class="event__offer-selector">
+            <input class="event__offer-checkbox  visually-hidden" id="event-offer-comfort-${point.id}" type="checkbox" name="event-offer-comfort" checked>
+            <label class="event__offer-label" for="event-offer-comfort-${point.id}">
+              <span class="event__offer-title">Switch to comfort</span>
+              &plus;&euro;&nbsp;
+              <span class="event__offer-price">80</span>
+            </label>
+          </div>
+          <div class="event__offer-selector">
+            <input class="event__offer-checkbox  visually-hidden" id="event-offer-meal-${point.id}" type="checkbox" name="event-offer-meal">
+            <label class="event__offer-label" for="event-offer-meal-${point.id}">
+              <span class="event__offer-title">Add meal</span>
+              &plus;&euro;&nbsp;
+              <span class="event__offer-price">15</span>
+            </label>
+          </div>
+          <div class="event__offer-selector">
+            <input class="event__offer-checkbox  visually-hidden" id="event-offer-seats-${point.id}" type="checkbox" name="event-offer-seats">
+            <label class="event__offer-label" for="event-offer-seats-${point.id}">
+              <span class="event__offer-title">Choose seats</span>
+              &plus;&euro;&nbsp;
+              <span class="event__offer-price">5</span>
+            </label>
+          </div>
+          <div class="event__offer-selector">
+            <input class="event__offer-checkbox  visually-hidden" id="event-offer-train-${point.id}" type="checkbox" name="event-offer-train">
+            <label class="event__offer-label" for="event-offer-train-${point.id}">
+              <span class="event__offer-title">Travel by train</span>
+              &plus;&euro;&nbsp;
+              <span class="event__offer-price">40</span>
+            </label>
+          </div>
         </div>
       </section>
       <section class="event__section  event__section--destination">
         <h3 class="event__section-title  event__section-title--destination">Destination</h3>
-        <p class="event__destination-description">${pointDestination.description}</p>
-        <div class="event__photos-container">
-          <div class="event__photos-tape">
-          ${createPicturesList}
-          </div>
-        </div>
+        <p class="event__destination-description">Chamonix-Mont-Blanc (usually shortened to Chamonix) is a resort area near the junction of France, Switzerland and Italy. At the base of Mont Blanc, the highest summit in the Alps, it's renowned for its skiing.</p>
       </section>
     </section>
   </form>
 </li>
-`);
+`;
 }
 
-export default class editPointView {
-  constructor(point, destinations, offers) {
-    this.point = point;
-    this.destinations = destinations;
-    this.offers = offers;
+export default class editFormView extends AbstractView {
+  #point = null;
+
+  constructor(point) {
+    super();
+    this.#point = point;
   }
 
-  getTemplate() {
-    return createNewPointTemplate(this.point, this.destinations, this.offers);
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
+  get template() {
+    return createEditPointTemplate(this.#point);
   }
 }
