+++ src/view/point-view.js
@@ -1,5 +1,5 @@
-import { createElement } from '../render.js';
-import { humanizeTravelDate, humanizeTimeFromTo, humanizeTravelTime } from '../utils.js';
+import AbstractView from '../framework/view/abstract-view.js';
+import { humanizeTravelDate, humanizeTimeFromTo, humanizeTravelTime } from '../utils/utils.js';
 
 const createOffersTemplate = (offerList) => offerList.map((offer) =>
   ` <li class="event__offer">
@@ -59,26 +59,29 @@ function createPointTemplate(point, destinations, offers) {
 `);
 }
 
-export default class PointView {
-  constructor(point, destinations, offers) {
-    this.point = point;
-    this.destinations = destinations;
-    this.offers = offers;
+export default class PointView extends AbstractView {
+  #point = null;
+  #destinations = null;
+  #offers = null;
+  #handleEditClick = null;
+
+  constructor({ point, destinations, offers, onEditClick }) {
+    super();
+    this.#point = point;
+    this.#destinations = destinations;
+    this.#offers = offers;
+    this.#handleEditClick = onEditClick;
+
+    this.element.querySelector('.event__rollup-btn')
+      .addEventListener('click', this.#editClickHander);
   }
 
-  getTemplate() {
-    return createPointTemplate(this.point, this.destinations, this.offers);
+  get template() {
+    return createPointTemplate(this.#point, this.#destinations, this.#offers);
   }
 
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
-  }
+  #editClickHander = (evt) => {
+    evt.preventDefault();
+    this.#handleEditClick();
+  };
 }
